+++ moai_app/lib/main.dart
@@ -2,7 +2,8 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:moai_app/services/huddle01/huddle_playground.dart';
+import 'package:moai_app/playground/huddle_playground.dart';
+import 'package:moai_app/playground/playground_launcher.dart';
 import 'package:moai_app/services/xmtp/xmtpinterface.dart';
 
 final navigatorKey = GlobalKey<NavigatorState>();
@@ -36,7 +37,7 @@ class MoaiApplication extends StatelessWidget {
           useMaterial3: true,
         ),
         navigatorKey: navigatorKey,
-        home: const HuddlePlayground(),
+        home: const PlaygroundLauncher(),
       ),
     );
   }