+++ moai_app/lib/playground/playground_launcher.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:moai_app/extensions/extensions.dart';
+import 'package:moai_app/playground/huddle_playground.dart';
 import 'package:moai_app/playground/wallet_playground.dart';
 
 class PlaygroundLauncher extends StatelessWidget {
@@ -6,7 +8,29 @@ class PlaygroundLauncher extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    // return const SingletPlayground();
-    return const WalletPlayground();
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Moai Playground'),
+        centerTitle: true,
+      ),
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          ElevatedButton(
+            onPressed: () {
+              Navigator.of(context).pushNewPage(const WalletPlayground());
+            },
+            child: Text('Wallet & XMTP Playground'),
+          ),
+          ElevatedButton(
+            onPressed: () {
+              Navigator.of(context).pushNewPage(const HuddlePlayground());
+            },
+            child: Text('Huddle01 Playground'),
+          )
+        ],
+      ).center(),
+    );
   }
 }